+++ client/src/views/profile/subviews/settings-subview.js
@@ -6,12 +6,12 @@ import AppContext from "context/app-context";
 import Button from "react-bootstrap/Button";
 import axios from "axios";
 import AvatarSelectionModal from "components/modal/avatar-selection-modal";
-import LoadingSpinner from "components/util/loading-spinner";
 import Swal from "sweetalert2";
 import swalReact from "sweetalert2-react-content";
 import {useHistory} from "react-router-dom";
 import ViewBox from "components/viewbox/view-box";
 import ActionButton from "components/app/action-button";
+import ComponentLoader from "components/app/component-loader";
 
 const SettingsSubview = ({reRouteTo}) => {
     const context = useContext(AppContext);
@@ -19,7 +19,7 @@ const SettingsSubview = ({reRouteTo}) => {
     const swalGenerator = swalReact(Swal);
     const [username, setUsername] = useState(context.user.data.username);
     const [avatar, setAvatar] = useState(context.user.data.avatar);
-    const [connectedAccounts, setConnectedAccounts] = useState(null);
+    const [connectedAccounts, setConnectedAccounts] = useState({data: [], loaded: false, error: false});
     const [modalOpened, setModalOpened] = useState(false);
     const onChangesSave = () => {
         let toastId = toastAwait("Saving changes...");
@@ -85,8 +85,9 @@ const SettingsSubview = ({reRouteTo}) => {
             if (res.status !== 200) {
                 return;
             }
-            setConnectedAccounts(res.data);
-        });
+            setConnectedAccounts({...connectedAccounts, data: res.data, loaded: true});
+        }).catch(() => setConnectedAccounts({...connectedAccounts, loaded: true, error: true}));
+        // eslint-disable-next-line
     }, []);
     if (!context.user.loggedIn) {
         return <React.Fragment>
@@ -99,15 +100,14 @@ const SettingsSubview = ({reRouteTo}) => {
         </React.Fragment>
     }
     const renderContent = () => {
-        if (connectedAccounts === null) {
-            return <LoadingSpinner/>
-        }
         return <React.Fragment>
-            <AvatarSelectionModal open={modalOpened} onAvatarModalClose={() => setModalOpened(false)}
-                                  connectedAccounts={connectedAccounts} onAvatarChoose={av => {
-                setModalOpened(false);
-                setAvatar(av);
-            }}/>
+            <ComponentLoader loaded={connectedAccounts.loaded} component={
+                <AvatarSelectionModal open={modalOpened} onAvatarModalClose={() => setModalOpened(false)}
+                                      connectedAccounts={connectedAccounts} onAvatarChoose={av => {
+                    setModalOpened(false);
+                    setAvatar(av);
+                }}/>
+            } loader={<React.Fragment/>}/>
             <Col xs={12} lg={6} className="order-lg-1 order-2">
                 <Form.Label className="mr-1 mt-lg-0 mt-2 text-black-60">Username</Form.Label>
                 <Form.Control style={{minHeight: 38, resize: "none"}} minLength="4" maxLength="20" rows="1" required
@@ -125,8 +125,10 @@ const SettingsSubview = ({reRouteTo}) => {
                 <Form.Label className="mr-1 text-black-60">Avatar</Form.Label>
                 <br/>
                 <img alt="avatar" src={avatar} className="img-fluid rounded-circle" width={100}/>
-                <Button variant="success" className="align-top mx-3 my-0"
-                        onClick={() => setModalOpened(true)}>Change</Button>
+                <ComponentLoader loaded={connectedAccounts.loaded}
+                                 component={<Button variant="success" className="align-top mx-3 my-0" onClick={() => setModalOpened(true)}>Change</Button>}
+                                 loader={<Button variant="success" disabled className="align-top mx-3 my-0">Loading</Button>}
+                />
             </Col>
             <Col xs={12} lg={6} className="order-3">
                 <Form.Label className="mr-1 mt-2 text-black-60">Email</Form.Label>